requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.2
Django==3.2.5
djangorestframework==3.12.4
idna==3.2
mysql-connector-python==8.0.25
protobuf==3.17.3
PyMySQL==1.0.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tabulate==0.8.9
urllib3==1.26.6
